Made a models package to separate the classes
