+++ backend/mental_backend/api/models/__init__.py
@@ -1,0 +1,2 @@
+from .journal import Journal
+from .note import Note
